# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Projeto de PDWA5",
        "description": "API do projeto de PDWA5",
        "contact": {
            "email": "admin@admin.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost/api",
            "description": "Endpoint da API."
        }
    ],
    "paths": {
...
        "/user": {
            "put": {
                "tags": [
                    "User"
                ],
                "summary": "Editar usuário.",
                "description": "Edita um usuário.",
                "operationId": "49d38ea7fe04e833b5be3c31258fa551",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "User Name"
                                    },
                                    "email": {
                                        "type": "string",
                                        "example": "email@example.com"
                                    },
                                    "pwd": {
                                        "type": "string",
                                        "example": "12345#qwert"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Usuário editado com sucesso.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "name": {
                                            "type": "string",
                                            "example": "User Name"
                                        },
                                        "email": {
                                            "type": "string",
                                            "example": "email@example.com"
                                        },
                                        "updated_at": {
                                            "type": "string",
                                            "example": "2022-11-14T00:06:03.000000Z"
                                        },
                                        "created_at": {
                                            "type": "string",
                                            "example": "2022-11-14T00:06:03.000000Z"
                                        },
                                        "id": {
                                            "type": "integer",
                                            "example": "1"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Ao falhar na autorização."
                    },
                    "409": {
                        "description": "Caso já exista um usuário com o mesmo nome e/ou e-mail."
                    },
                    "422": {
                        "description": "Ao falhar em uma validação.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "name": {
                                                "type": "array",
                                                "items": {}
                                            },
                                            "email": {
                                                "type": "array",
                                                "items": {}
                                            },
                                            "pwd": {
                                                "type": "array",
                                                "items": {}
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "User"
                ],
                "summary": "Criar usuário.",
                "description": "Cria um novo usuário.",
                "operationId": "fdaf6e20ea63c9e43bacb0c1f25012d5",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "User Name"
                                    },
                                    "email": {
                                        "type": "string",
                                        "example": "email@example.com"
                                    },
                                    "pwd": {
                                        "type": "string",
                                        "example": "12345#qwert"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Usuário criado com sucesso.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "name": {
                                            "type": "string",
                                            "example": "User Name"
                                        },
                                        "email": {
                                            "type": "string",
                                            "example": "email@example.com"
                                        },
                                        "updated_at": {
                                            "type": "string",
                                            "example": "2022-11-14T00:06:03.000000Z"
                                        },
                                        "created_at": {
                                            "type": "string",
                                            "example": "2022-11-14T00:06:03.000000Z"
                                        },
                                        "id": {
                                            "type": "integer",
                                            "example": "1"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "409": {
                        "description": "Caso já exista um usuário com o mesmo nome e/ou e-mail."
                    },
                    "422": {
                        "description": "Ao falhar em uma validação.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "name": {
                                                "type": "array",
                                                "items": {}
                                            },
                                            "email": {
                                                "type": "array",
                                                "items": {}
                                            },
                                            "pwd": {
                                                "type": "array",
                                                "items": {}
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "User"
                ],
                "summary": "Remover usuário.",
                "description": "Remove o usuário.",
                "operationId": "a14bc5aa7e52b6b25b79f173f6639099",
                "responses": {
                    "200": {
                        "description": "Usuário removido com sucesso."
                    },
                    "401": {
                        "description": "Ao falhar na autorização."
                    },
                    "404": {
                        "description": "Caso o usuário não exista."
                    }
                }
            }
        }
    }
}